Support view column comments for DeltaLake
