+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeConnectorTest.java
@@ -151,9 +151,6 @@ public class TestDeltaLakeConnectorTest
             case SUPPORTS_SET_COLUMN_TYPE:
                 return false;
 
-            case SUPPORTS_COMMENT_ON_VIEW_COLUMN:
-                return false;
-
             case SUPPORTS_CREATE_MATERIALIZED_VIEW:
                 return false;
